Create simple HTTP-logging middleware
This creates a simple middleware that logs requests received to the
given logging API endpoint in the format: `"GET http://example.com"`.

The actual Readme.io payload format is more complicated than this. We
will add support for it in a follow-up commit.

When comparing Rack mock responses to real ones, we needed to convert to
the raw status, headers, body triple ourselves since there isn't a
built-in method to do so and we needed a way to compare against the real
triple.

_Note that Rack::MockResponse converts the array body into a
single string for ease of testing so we had to convert back to an array
as part of destructuring._