Send valid payloads to Readme API
This sends a valid hardcoded payload (in the form of the `data.json` file)
to the Readme API with all the proper headers and authentication.

The middleware now takes an API key during initialization and no longer
needs an endpoint which has become a constant.

The hard-coded data we wrote is valid according to [HAR
1.2](http://www.softwareishard.com/blog/har-12-spec). We noticed that
there are some required fields from the spec that are neither mentioned
by the API nor used in the nodejs implementation. We decided to include
them for validity sake and at some point the backend might want to fully
support the HAR 1.2 spec.

The added fields are:

* `entries.cache`
* `entries.timing`
* `entries[].request.cookies`
* `entries[].request.headerSize`
* `entries[].request.bodySize`
* `entries[].response.cookies`
* `entries[].response.headerSize`
* `entries[].response.bodySize`
* `entries[].response.redirectURL`
* `entries[].response.httpVersion`

Some of these are empty arrays or objects but the keys are required to
be present.

As part of our development process, we introduced json-schema to check
the generated JSON we use (for now the static JSON file). We wrote a
schema file for the Readme.io-specific structure but were able to re-use
an [existing implementation for the HAR
format](https://github.com/ahmadnassri/node-har-schema).