Move JSON generation out of the middleware
This commit is a refactor that moves JSON generation out of the
middleware and in to two new classes `Readme::Har` and
`Readme::Payload`.

We added a new custom matcher for the JSON schema using the [json-schema
gem] based on [this blog post]. We were unable to use an [existing gem]
that contained a similar RSpec matcher because it is built on top of the
[json_schema gem] (note the underscore rather than dash) that seems to
be having issues when expanding references to other local schema files.

[this blog post]: https://github.com/thoughtbot/json_matchers
[json-schema gem]: https://github.com/ruby-json-schema/json-schema
[existing gem]: https://github.com/thoughtbot/json_matchers
[json_schema gem]: https://github.com/brandur/json_schema