Generate real HAR from request and response
Previously, we were using hard-coded values for the HAR. Now we
correctly derive a HAR from the Rack env (request) and response. This is
currently only implemented and tested for the "happy path" of POST
requests with both request and response bodies. Follow-up commits will
broaden the scope of handled request types.

The HAR spec was designed to describe requests made from a client so
there are some keys that are nonsensical in the context of a middleware.
We've either left those out or set the appropriate "we don't know" value
(e.g. -1 for response header size).

We noticed that `Rack::Test` doesn't set the HTTP_VERSION header even
though actual rack servers set it. To get around this, we added another
little middleware in tests that sets that header to `"HTTP/1.2"`.