Validate user info block configuration
This commit adds some validation around user info block required by the
middleware.

One check runs at application boot and ensures that the block is not
nil.

The other checks occur at run time and ensure that if the block is not
configured properly, the application will not crash and log an error
instead. The request to the Readme API will be aborted and the response
will be passed on through the rest of the middleware stack.