Add support for empty response bodies
Some responses have no response body (e.g. an HTTP 204 response).
`Rack::Response` handles this by returning `nil` when calling `#body`.

The HAR spec mentions that `response.content.text` key is optional.
However, it _does_ require the `response.content.mimeType` key. Based on
discussion with the Readme.io team, we decided to send an empty string in
this spot when there is no request body.

The HAR spec is ambiguous about the `response.content.size` key. It
mentions that it should be 0 for 304 responses (another scenario with no
body) and -1 if unknown. I chose to set 0 for responses without a body
since we do know the length and the spec specifically says to use 0
in some instances of no-body responses.

While testing this code, I found a bug where attempting to read the body
more than once failed since the IO object had not been rewound.

While moving some of the existing JSON code around, I realized that some
of it's code paths were untested so I added a test around invalid
payloads.