Perform API request to ReadMe in a new thread
This commit updates the behavior of the middleware to perform
non-blocking requests to the ReadMe API. When the request queue is full,
the request to ReadMe is performed in a new thread. The application's
response is passed through straight away.

In tests, we had to stub the `Thread` class to ensure tests are run
synchronously and deterministically.