Log an error if we abort the request to ReadMe
When something goes wrong in the middleware, we rescue and just return
to the next middleware in the chain.

This commit adds some logging in that case so that users can better
understand why some requests aren't being submitted to the ReadMe API.
This might be due to a bug in this middleware (in which case this error
will allow users to create better bug reports), or it might be due to a
malformed body (in which case surfacing that to the user is helpful).